runtime/requirements.txt
litellm>=1.40.0
aiogram>=3.4.0
apscheduler>=3.10.0
pyyaml>=6.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
